Database Creation
This is going to need to be redone. Game id isn't going to be unique so at best we have a bunch of composite keys, but creating a unique table to hold a primary game id also creates redundant information. I am sure something will come to me if I sleep on it. Pushing this for the users table so I can start work on the login system.