Login session carries between pages
Added some buttons and redirects to go to other pages without manually typing in the URL or relying on the back buttons. Emails are no longer unique which is /probably/ fine if I ask for usernames and return to their email without specifying email it is, but I will need to reassess later. Perhaps a check for unique email server-side can do the trick. Also set this up with session instead of cookies for security purposes.